requirements.txt
torch
torchvision
transformers
pillow
chromadb
gradio
numpy
